# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "sooandcarrots-api",
        "description": "This is a simple example of a OpenAPI definition",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://sooandcarrots/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "60e6ec72fbb2dd5f8977b3623a5bd2f6",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/user/pagination": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get 10 users at a time",
                "description": "Returns list of 10 users ",
                "operationId": "cdbcf58c0999e9f8c306bb1484e54786",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/users/likes": {
            "post": {
                "tags": [
                    "User Likes"
                ],
                "summary": "Update user likes or dislikes",
                "description": "Allows updating the likes or dislikes count for a specified user based on the action provided.",
                "operationId": "updateUserPreferences",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The ID of the user whose preferences are to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The action to perform (like or dislike)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "action"
                                ],
                                "properties": {
                                    "action": {
                                        "description": "Action to perform on user's preferences",
                                        "type": "string",
                                        "enum": [
                                            "like",
                                            "dislike"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preferences updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User preferences updated successfully."
                                        },
                                        "user": {
                                            "properties": {
                                                "likes": {
                                                    "description": "Total number of likes",
                                                    "type": "integer"
                                                },
                                                "dislikes": {
                                                    "description": "Total number of dislikes",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid action type provided."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to update user preferences due to an unexpected error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user-likes": {
            "get": {
                "tags": [
                    "User Likes"
                ],
                "summary": "Get all user likes",
                "description": "Retrieves a list of all user likes.",
                "operationId": "listUserLikes",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "user_id": {
                                                "description": "ID of the user",
                                                "type": "integer"
                                            },
                                            "likes": {
                                                "description": "Number of likes",
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No likes found"
                    }
                }
            }
        }
    }
}